Update main.cpp
Problem solved. The issue was caused by a texture upload that occurred too early. This resulted in the first read image's "fill" flag (which basically is used to indicate that "yes, image X has been inserted into an atlas", so an accidental second insertion isn't attempted) to be set when it shouldn't have been set yet.